Add files via upload
Added the main NSintegrator.ipynb wich allows to solves the TOV equations for static neutron stars described by an EOS that includes a QCD and a vacuum energy phase transition. It also allows to compute the combined tidal deformability of a NSs binary.
Added the T4.ipynb which compute the inspirals for a NS binary using post-Newtonian description using the Taylor T4 formalism with tidal contributions discussed in https://arxiv.org/abs/1301.3555v2 (and references therein).